fixes the selection reset bug that happens after every refresh
The periodic update of the port bindings was causing a weird bug where the selection in the view was reset to the 0 position, but the focus stayed at the correct position.

Apparently for table views there is a 'focus' which is rendered as a blue bar, and a 'selection' which is rendered as a narrow border.

The problem was caused because the `setItems` method of the ObservableList changes the 'selection' index to 0, but leaves the 'focus' where it was before if possible.

The workaround is to create a snapshot before every update, and calculate the additions, and removals between current, and previous state, and then use two functions `removeAll` and `addAll` to update the list. This updates both 'selection', and 'focus' as I expect.