uses 'compile' instead of 'runtime' as configuration for dependencies
Setting the configuration to 'runtime' from 'implementation' solved the problem that 'jfxJar' did not copy the dependencies. Now, IntelliJ would not compile the code any longer since it was missing dependencies at compile time. Setting the configuration to 'compile' seems to solve both problems. I assume that the plugin is a bit older and has not caught on the Gradle standard changing from 'compile' to 'implementation'.